+++ src/llm_lab/verifier/__init__.py
@@ -1,0 +1,4 @@
+from .interface import Verifier
+from .docker_verifier import DockerVerifier
+
+__all__ = ["Verifier", "DockerVerifier"]